Add prototype sequence tester
Adds a basic sequence tester pattern drawing off the UVM style pattern
of sequences, drivers, and monitors.

For now, we simply enable the definitation of input and output sequences
as a Python Sequence of values.  Drivers and Monitors are functions that
take a reference to the tester and the current input/output value.
Inside the function they can lower the value into an arbitrary sequence
of test actions.

There's definitely more advanced patterns we should consider, such as
monitors being based purely on dynamic values (e.g. a monitor could just
watch a bus and make sure transactions are valid using invariants, so it
wouldn't need to be associated with an output sequence).  There also may
be variable length sequences (e.g. every 4 input values corresponds to 1
output value).

However, this basic pattern is intended to target the testing of the
garnet PE core/tile (thus the test/example is a simple instace of this
pattern).  The goal is to reuse the same input/output sequence for both
the core and tile.  This is done by "lifting" the sequences to the tile
level by defining a new driver (in this case the monitor is the same,
but for a more complex example, you may need to do some different logic,
e.g. read the right connection box value based on a randomized
configuration).

I'm open to changes to the interface (especially with the above examples
or others in mind), this was simply something basic I came up for the
target use case.