Add var poke test for signed values
It turns out fault converts signed values in poke/expect to unsigned
values (see process_signed_values).  This is because verilator uses
unsigned C types for the ports (even though in magma it may be an SInt).

I also updated the var test's interface to be more consistent by using
BitVector and SIntVector instead of m.Bits (since this is testing I
think it makes sense to use the hwtypes versus magma types).  I had to
make a small change to make_value to support that (no conversion is
necessary since we're already using an appropriate BitVector type).

I remove the signed specific cdoe for the var poke since it's not needed
(since the poke value will be converted to unsigned by
process_signed_value, we should store it as unsigned, the same way that
verilator does it).

This is similar to how the system verilog test bench works (magma/coreir
does not generate the signed prefix for m.SInt types, so they are
treated as unsigned in the design and test benches, so we convert signed
python values to unsigned when generating code).