Add support for call statements
Separates make_call into two functions: one for an expression and one
for a statement.

Reworks the backend call logic to treat it as an expression rather than
a value (also adds a test that uses it in an expression).

Call statements are useful for making calls to a model that don't
return a value (e.g. just manipulates the state)