Add ReadyValidTester for I/O sequences
For an initial implementation we can actually create a synthesizable
test bench to feed input sequences and check output sequences.  This is
likely to become an issue when scaling to large sequences (since we have
to generate each sequence element as a constant fed into a mux).  To
handle this, I think we can change the sequence mux to read data from a
file (or a memory that has its contents loaded from a file).